+++ .gitignore
@@ -2,3 +2,4 @@ node_modules
 # Keep environment variables out of version control
 .env
 images
+dist
