+++ src/app.js
@@ -5,8 +5,8 @@ import session from 'express-session';
 import morgan from 'morgan';
 import passport from 'passport';
 import path from 'path';
-import env from '../configs/';
-import passportSession from '../configs/passport';
+import env from './configs/';
+import passportSession from './configs/passport';
 import BannerController from './controllers/BannerController';
 import KeywordController from './controllers/KeywordController';
 import PostController from './controllers/PostController';
