+++ src/configs/algolia/index.js
