+++ src/controllers/BannerController.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import * as Auth from '../../middleware/auth';
+import * as Auth from '../middleware/auth';
 import * as BannerService from '../services/BannerService';
 
 const router = express.Router();
